Add logging to WebHttpConnector

+++ src/Moryx.Tools.Wcf/Client/Connector/WebHttpServiceConnectorBase.cs
@@ -8,6 +8,8 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using Moryx.Communication;
+using Moryx.Container;
+using Moryx.Logging;
 using Newtonsoft.Json;
 
 namespace Moryx.Tools.Wcf
@@ -24,6 +26,11 @@ namespace Moryx.Tools.Wcf
         /// Name of the service interface to connect to
         /// </summary>
         public abstract string ServiceName { get; }
+
+        /// <summary>
+        /// Logger of the connector
+        /// </summary>
+        protected IModuleLogger Logger { get; }
 
         /// <summary>
         /// Gets the current client version of the client
@@ -45,20 +52,23 @@ namespace Moryx.Tools.Wcf
                 AvailabilityChanged?.Invoke(this, EventArgs.Empty);
             }
         }
+
         /// <summary>
         /// Create connector using the client factories version service
         /// </summary>
-        protected WebHttpServiceConnectorBase(IWcfClientFactory clientFactory)
+        protected WebHttpServiceConnectorBase(IWcfClientFactory clientFactory, IModuleLogger logger)
         {
-            _endpointService = ((BaseWcfClientFactory)clientFactory).VersionService;
+            var baseFactory = (BaseWcfClientFactory)clientFactory;
+            _endpointService = baseFactory.VersionService;
+            Logger = logger;
         }
-
         /// <summary>
-        /// Create connector with bare connection settings
+        /// Create connector with bare connection settings and Logger
         /// </summary>
-        protected WebHttpServiceConnectorBase(string host, int port, IProxyConfig proxyConfig)
+        protected WebHttpServiceConnectorBase(string host, int port, IProxyConfig proxyConfig, IModuleLogger logger)
         {
             _endpointService = new VersionServiceManager(proxyConfig, host, port);
+            Logger = logger;
         }
 
         /// <inheritdoc />
@@ -66,6 +76,7 @@ namespace Moryx.Tools.Wcf
         {
             TryFetchEndpoint();
         }
+
 
         private void TryFetchEndpoint()
         {
@@ -78,6 +89,7 @@ namespace Moryx.Tools.Wcf
             //Try again or dispose old client
             if (resp.Status != TaskStatus.RanToCompletion || resp.Result.Length == 0)
             {
+                Logger.Log(LogLevel.Warning, "Failed to read endpoints.");
                 await CallbackAndTryFetch(ConnectionState.FailedTry);
                 return;
             }
@@ -86,6 +98,8 @@ namespace Moryx.Tools.Wcf
             var endpoint = resp.Result.FirstOrDefault(e => e.Binding == ServiceBindingType.WebHttp);
             if (endpoint == null || string.IsNullOrEmpty(endpoint.Address))
             {
+                Logger.Log(LogLevel.Error, "Endpoint for {0} has wrong binding or empty address: {1}-{2}", 
+                    ServiceName, endpoint?.Binding, endpoint?.Address);
                 await CallbackAndTryFetch(ConnectionState.FailedTry);
                 return;
             }
@@ -97,7 +111,7 @@ namespace Moryx.Tools.Wcf
             if (serverVersion.Major == clientVersion.Major & serverVersion >= clientVersion)
             {
                 // Create new base address client
-                HttpClient = new HttpClient {BaseAddress = new Uri(endpoint.Address)};
+                HttpClient = new HttpClient { BaseAddress = new Uri(endpoint.Address) };
 
                 IsAvailable = true;
 
@@ -105,6 +119,8 @@ namespace Moryx.Tools.Wcf
             }
             else
             {
+                Logger.Log(LogLevel.Error, "Version mismatch: Client: {0} - Server: {1}", 
+                    clientVersion, serverVersion);
                 await CallbackAndTryFetch(ConnectionState.VersionMissmatch);
             }
         }
@@ -142,6 +158,9 @@ namespace Moryx.Tools.Wcf
         /// </summary>
         protected async Task<T> GetAsync<T>(string url)
         {
+            if(!IsAvailable)
+                throw new InvalidOperationException("Client not available!");
+
             var response = await HttpClient.GetStringAsync(url);
             return JsonConvert.DeserializeObject<T>(response);
         }
@@ -151,6 +170,9 @@ namespace Moryx.Tools.Wcf
         /// </summary>
         protected async Task<T> PostAsync<T>(string url, object payload)
         {
+            if (!IsAvailable)
+                throw new InvalidOperationException("Client not available!");
+
             var payloadString = string.Empty;
             if (payload != null)
                 payloadString = JsonConvert.SerializeObject(payload);
@@ -165,6 +187,9 @@ namespace Moryx.Tools.Wcf
         /// </summary>
         protected async Task<T> PutAsync<T>(string url, object payload)
         {
+            if (!IsAvailable)
+                throw new InvalidOperationException("Client not available!");
+
             var payloadString = string.Empty;
             if (payload != null)
                 payloadString = JsonConvert.SerializeObject(payload);
@@ -179,6 +204,9 @@ namespace Moryx.Tools.Wcf
         /// </summary>
         protected async Task<bool> DeleteAsync(string url)
         {
+            if (!IsAvailable)
+                throw new InvalidOperationException("Client not available!");
+
             var response = await HttpClient.DeleteAsync(url);
             return response.StatusCode == HttpStatusCode.OK;
         }
